==> animal_farm/ghg/STEP2_animal_ghg_manure_n2o.Rmd <==
---
title: "Calculating manure emissions rates for all livestock"
date: "7/1/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---

Author: Juliette Verstaen & MRF

Objectives: We want to calculate an CO2-eq emissions factor for each animal/product combination. There are two types of emissions, CH4 and N2O, and three paths that manure can follow that have different emissions: managed (then applied to soil), applied directly to soils, and left on fields.

We use GLEAM data to identify the proportion of manure that goes down each path for each country and animal system.  We also use GLEAM data to estimate emissions from managed manure based on what systems countries tend to use to manage their manure for given animals and rearing systems.  We use FAO data to estimate emissions from manure directly applied to soils and manure that is left on fields.

## Setup
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(here)
library(stringr)
library(doParallel)
library(janitor)
library(countrycode)
library(stringr) 
library(dplyr)
source(here("_workflow/common.R"))
source(here("_workflow/common_spatial.R"))

un <- read_csv(here("_spatial/_output/UNSD_Methodology.csv")) %>% 
  dplyr::select(iso3c, Global_Name, Region_Name, Sub_region_Name, Intermediate_Region_Name)

N20_to_CO2 <- 298

```

## Emissions rates for manure left on pasture and spread on fields (from FAO)
For manure left on field and manure applied to soils we will use the FAO values calculated for each animal (was going to do by animal and country, but the values were basically the same across all countries, so we went with that).  


#### left on pasture
```{r}

lop <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_leftonpasture_2-18-2021.csv") %>%
  filter(Element == "Implied emission factor for N2O (Manure on pasture)") %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value)

lop$iso3c <- countrycode(as.character(lop$Area), origin="country.name", destination = "iso3c")
lop <- lop %>%
  mutate(iso3c = ifelse(Area == "French Guyana", "GUF", iso3c))
tmp <- lop %>%
  pivot_wider(names_from=Item, values_from = Value) %>% data.frame()
summary(tmp)

# was going to do something more complicated, but these values are essentially the same for all countries

lop_final <- data.frame(animal = c("buffaloes", "chickens", "cows", "goats", "pigs", "sheep"),
                        adj_value = c(0.024, 0.024, 0.024, 0.0143, 0, 0.0143)) %>%
   mutate(N20_to_N_leftonfields = adj_value*44/28) %>%  # converts from N20-N to N20
  select(-adj_value)

```


#### spread on fields
```{r}
sof <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_appliedtosoils_2-18-2021.csv") %>%
  filter(Element == "Implied emission factor for N2O (Manure applied)") %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value)

sof$iso3c <- countrycode(as.character(sof$Area), origin="country.name", destination = "iso3c")
sof <- sof %>%
  mutate(iso3c = ifelse(Area == "French Guyana", "GUF", iso3c))
summary(sof)

tmp <- sof %>%
  pivot_wider(names_from=Item, values_from = Value) %>% data.frame()
summary(tmp)

sof_final <- data.frame(animal = c("buffaloes", "chickens", "cows", "goats", "pigs", "sheep"),
                        adj_value = c(0.0143, 0.0143, 0.0143, 0.0143, 0.0143, 0.0143)) %>%
   mutate(N20_to_N_spreadonfields = adj_value*44/28) %>%  # converts from N20-N to N20
  select(-adj_value)


```


## Emissions rates for managed manure (from GLEAM)
This is calculated based on the type of manure management systems used in each country for each animal.  

Here we calculate several things:

- average N20 to N conversion for each animal system and country based on the weighted average of the emissions for each manure management system.
- proportion of manure managed
- leftover proportion of N after manure management


The direct and indirect emissions are summed:
*4.4.2 Direct N2O Emissions*
N2O_direct = (44/28) * number animals * excretion rate * Esum(emission factor for direct n2o for mms X * fraction manure going to mms X)

     = (44/28) * head_count * excr_rate * Esum(ef_dir * frac_mms)

emission factor for direct n2o for mms X = manure_n2o_em_factors.csv
ef_dir = gleam_table_4.15_ef_dir

*4.4.3 Indirect N2O emissions: volatization*
N2O = (44/28) * emission factor for n2o em from N volt * number animals * excretion rate * Esum(fraction manure going to mms X* (Percent manure that volatilizes in mms X)/100)

    = (44/28) * ef_gas * head_count * excr_rate * Esum(frac_mms * (frac_gas)/100)

ef_gas = 0.01
frac_gas = gleam_table_4.16_frac_gas
frac_mms = in mms_specific_percentages.csv

*4.4.4 Indirect N2O emissions: leaching*
N2O = (44/28) * emission factor for n2o em from N leaching * number animals * excretion rate * Esum(fraction manure going to mms X* (Percent manure that leaches in mms X)/100)

    = (44/28) * ef_leach * head_count * excr_rate * Esum(frac_mms* (frac_leach)/100)

ef_leach = 0.0075
frac_leach = gleam_table_4.16_frac_leach
frac_mms = in mms_specific_percentages.csv

```{r}

gleam_tables <- read_csv(here("animal_farm/ghg/data/compiled_gleam_tables.csv"))

 managed_manure <- gleam_tables %>% 
      unique() %>%
      filter(!(mms_specific %in% c("Burned for fuel", "Pasture/Range/Paddock", "Daily spread"))) %>% # different paths that are analyzed later on in the script
      rowwise() %>%
      mutate(mms_direct = ef_direct * to_mms_percentage/100 * 44/28) %>%  # units: kg N2O - N kg/N, 44/28 converts from N20-N
      mutate(mms_indirect_leach = 0.0075 * percentage_leach/100 * to_mms_percentage/100 * 44/28) %>%     # 0.0075 is default of GLEAM
      mutate(mms_indirect_vol = 0.01 * frac_gas/100 * to_mms_percentage/100 * 44/28) %>%  # 0.01 default of GLEAM
      mutate(mms_total = mms_direct + mms_indirect_leach + mms_indirect_vol) %>%
      mutate(loss_N_prop = to_mms_percentage/100 * (percentage_leach/100 + frac_gas/100)) %>%
      group_by(iso3c, country, animal, product, production_system) %>% 
      dplyr::summarise(N_to_N20_managed_all_N = sum(mms_total, na.rm = TRUE),
                       loss_N_managed_prop = sum(loss_N_prop, na.rm=TRUE),
                       managed_manure_prop = sum(to_mms_percentage, na.rm=TRUE)/100) %>%
  rowwise() %>%
      dplyr::mutate(N_to_N20_managed = N_to_N20_managed_all_N/managed_manure_prop) %>% 
  select(-N_to_N20_managed_all_N) %>%
      ungroup() 

```

Now, we gapfill these data using a variety of methods.
```{r}
all_cats <- expand.grid(iso3c = un$iso3c, 
            animal = unique(paste(managed_manure$animal, managed_manure$product, managed_manure$production_system, sep="_"))) %>%
  separate(animal, into=c("animal", "product", "production_system"), sep="_") %>%
  filter(!is.na(iso3c))

managed_manure_gf <- left_join(all_cats, managed_manure, by = c("iso3c", "animal", "product", "production_system")) 
summary(managed_manure_gf)

gf_data_bypigs <- data.frame(animal = "pigs", product = "meat", production_system="backyard", loss_N_managed_prop = 0.4285, 
                      managed_manure_prop= 0.9, N_to_N20_managed=0.014) 
gf_data_indpigs <- data.frame(animal = "pigs", product = "meat", production_system=c("industrial", "intermediate"),
                                loss_N_managed_prop = 0.4445, managed_manure_prop= 1, N_to_N20_managed=0.022) 
gf_data_indchickens <- data.frame(animal = "chickens", product = c("meat", "eggs"), production_system=c("industrial"),            
                                  loss_N_managed_prop = 0.46, managed_manure_prop= 1, N_to_N20_managed=0.01)

## gapfill industrial chicken data  
missing_chickens<- filter(managed_manure_gf, animal=="chickens" & production_system == "industrial" & (is.na(N_to_N20_managed)|N_to_N20_managed==0)) %>%
  select(iso3c, animal, product, production_system, country)

gf_data_indchickens_add <- left_join(missing_chickens, gf_data_indchickens)

missing_ind_pigs<- filter(managed_manure_gf, animal=="pigs" & production_system %in% c("industrial", "intermediate") & (is.na(N_to_N20_managed)|N_to_N20_managed==0)) %>%
  select(iso3c, animal, product, production_system, country)

gf_data_indpigs_add <- left_join(missing_ind_pigs, gf_data_indpigs)

missing_by_pigs<- filter(managed_manure_gf, animal=="pigs" & production_system %in% c("backyard") & (is.na(N_to_N20_managed)|N_to_N20_managed==0)) %>%
  select(iso3c, animal, product, production_system, country)

gf_data_bypigs_add <- left_join(missing_by_pigs, gf_data_bypigs)

## make adjustments to main dataframe
managed_manure_gf <- managed_manure_gf %>%
  filter(!(animal=="chickens" & production_system == "industrial" & (is.na(N_to_N20_managed)|N_to_N20_managed==0))) %>%
  filter(!(animal=="pigs" & production_system %in% c("industrial", "intermediate") & (is.na(N_to_N20_managed)|N_to_N20_managed==0))) %>%
  filter(!(animal=="pigs" & production_system %in% c("backyard") & (is.na(N_to_N20_managed)|N_to_N20_managed==0))) %>%
  rbind(gf_data_indchickens_add) %>%
  rbind(gf_data_indpigs_add) %>%
  rbind(gf_data_bypigs_add) %>%
  filter(!(production_system == "feedlot" & is.na(N_to_N20_managed))) %>%
  rowwise() %>%
  mutate(loss_N_managed_prop = ifelse(production_system == "grassland", 0, loss_N_managed_prop)) %>%
  mutate(managed_manure_prop = ifelse(production_system == "grassland", 0, managed_manure_prop)) %>%
  mutate(N_to_N20_managed = ifelse(production_system == "grassland", 0, N_to_N20_managed)) %>%
  mutate(loss_N_managed_prop = ifelse(product == "eggs&meat", 0, loss_N_managed_prop)) %>%
  mutate(managed_manure_prop = ifelse(product == "eggs&meat", 0, managed_manure_prop)) %>%
  mutate(N_to_N20_managed = ifelse(product == "eggs&meat", 0, N_to_N20_managed)) %>%
  mutate(loss_N_managed_prop = ifelse(is.na(loss_N_managed_prop), 0, loss_N_managed_prop)) %>%
  mutate(managed_manure_prop = ifelse(is.na(managed_manure_prop), 0, managed_manure_prop)) %>%
  mutate(N_to_N20_managed = ifelse(is.na(N_to_N20_managed), 0, N_to_N20_managed)) %>% 
  ungroup()

summary(managed_manure_gf)
tmp <- filter(managed_manure_gf, is.na(N_to_N20_managed))
tmp <- filter(managed_manure_gf, animal=="cows" & production_system=="mixed")

```

## determine proportion of manure going down other paths
There are several paths manure can take: 1. Managed and then spread, 2. Directly spread with no management, 3. Burned, 4. left on field.  Here we use GLEAM data to determine the proportion of manure that goes down each path for each country and animal system.

Currently, we do not include burned category.

#### Burned
```{r}

 managed_manure_burned <- gleam_tables %>% 
      unique() %>%
      filter(mms_specific %in% c("Burned for fuel")) %>%
   select(iso3c, animal, product, production_system, burned_prop=to_mms_percentage) %>%
   mutate(burned_prop = burned_prop/100)

# if there is no value we assume 0% burned 
 managed_manure_all <- left_join(managed_manure_gf, managed_manure_burned, by=c("iso3c", "animal", "product", "production_system")) %>%
   mutate(burned_prop = ifelse(is.na(burned_prop), 0, burned_prop))

```
 
 
##### Left on field

```{r}

  managed_manure_lof <- gleam_tables %>% 
      unique() %>%
      filter(mms_specific %in% c("Pasture/Range/Paddock")) %>%
       select(iso3c, animal, product, production_system, left_on_field_prop=to_mms_percentage) %>%
   mutate(left_on_field_prop = left_on_field_prop/100)   
  
# join with managed manure data  
manure_all <- managed_manure_all %>%
    left_join(managed_manure_lof, by=c("iso3c", "animal", "product", "production_system")) %>%
    rowwise() %>%
    mutate(left_on_field_prop = ifelse(production_system == "grassland", 1, left_on_field_prop)) %>%
    mutate(left_on_field_prop = ifelse(product == "eggs&meat", 0.5, left_on_field_prop)) %>%
    mutate(left_on_field_prop = ifelse(animal=="pigs" & production_system=="backyard", 0.05, left_on_field_prop)) %>%
    mutate(left_on_field_prop = ifelse(animal=="pigs" & production_system=="intermediate", 0, left_on_field_prop)) %>%
    mutate(left_on_field_prop = ifelse(animal=="pigs" & production_system=="industrial", 0, left_on_field_prop)) %>%
    mutate(left_on_field_prop = ifelse(animal=="chickens" & production_system=="industrial", 0, left_on_field_prop)) %>%
    mutate(left_on_field_prop = ifelse(is.na(left_on_field_prop), 0, left_on_field_prop))
  
summary(manure_all)
tmp <- filter(manure_all, is.na(left_on_field_prop))
table(tmp$animal, tmp$product, tmp$production_system)


```


##### Spread on soil

```{r}

  managed_manure_sof <- gleam_tables %>% 
      unique() %>%
      filter(mms_specific %in% c("Daily spread")) %>%
       select(iso3c, animal, product, production_system, spread_on_field_prop=to_mms_percentage) %>%
   mutate(spread_on_field_prop = spread_on_field_prop/100)

  filter(managed_manure_sof, iso3c=="SDN")  

  # join with other manure data
  
  manure_all <- manure_all %>%
    left_join(managed_manure_sof, by=c("iso3c", "animal", "product", "production_system")) %>%
    rowwise() %>%
    mutate(spread_on_field_prop = ifelse(production_system == "grassland", 0, spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(product == "eggs&meat", 0.5, spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(animal=="pigs" & production_system=="backyard", 0.05, spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(animal=="pigs" & production_system=="intermediate", 0, spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(animal=="pigs" & production_system=="industrial", 0, spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(animal=="chickens" & production_system=="industrial", 0, spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(is.na(spread_on_field_prop), 1-(managed_manure_prop + burned_prop + left_on_field_prop), spread_on_field_prop)) %>%
    mutate(spread_on_field_prop = ifelse(spread_on_field_prop < 1, 1-(managed_manure_prop + burned_prop + left_on_field_prop), spread_on_field_prop)) %>%
    rowwise() %>%
    mutate(total_prop = managed_manure_prop + burned_prop + left_on_field_prop + spread_on_field_prop)

  summary(manure_all)
# should be no NA values and total_prop should equal one for everything
```


##### Combine with emission rates from FAO
These values were extracted from FAO (above in code).

```{r}
manure_all_em <- manure_all %>%
  left_join(lop_final, by = "animal") %>%
  left_join(sof_final, by = "animal") %>%
  rename(system = production_system)

summary(manure_all_em)

```

## Calculate emissions and map

```{r}

map_df_list <- list.files(file.path(prep, "animal_farm/farm"), pattern = "location_df",  full = TRUE)

# n_exc <- read_csv(here("animal_farm/ghg/data/nitrogen_excretion_rates.csv")) %>% 
#   select(-ID_0, -Country)

n_exc <- read_csv(here::here("animal_farm/nutrient/data/excretion_rate_gf.csv"))

c <- read_csv(file.path(prep, "animal_farm/farm/pigs_industrial_meat_location_df.csv"))

registerDoParallel(6)   
foreach(file = map_df_list)  %dopar%  {

     # file <- map_df_list[15]
     
    ## grab the names of all the parts from the file
    animal_name <- str_split(str_split(file, pattern = "/")[[1]][11], pattern = "_")[[1]][1]
    system_name <- str_split(str_split(file, pattern = "/")[[1]][11], pattern = "_")[[1]][2]
    product_name <- str_split(str_split(file, pattern = "/")[[1]][11], pattern = "_")[[1]][3]

    ### each animal group has slightly different map_df, so we need to specifu the column types for each
    
    
    if(animal_name == "chickens"){
      col_type = c("ddnnnccnnccncnnc")
      
      }else{
      if(animal_name %in% c("goats", "sheep", "cows", "buffaloes")) {    
      col_type = c("ddccccn")
        
      }else{
        ## this is pigs
        (col_type = c("ddcccnnnnncc")) } } 
    
    
    ## get the head count map:
    region_counts <- read_csv(file, col_types = col_type) %>% 
      select(x, y, iso3c, current = contains("current")) %>% 
      mutate(animal = animal_name,
             system = system_name,
             product = product_name)
  
    ## add on the country/animal specific excretion rates
    region_count_emissions <- left_join(region_counts, n_exc, by = c("iso3c", "animal", "product"))
  
    if(dim(region_count_emissions)[1] != 9331200){cat("error in count after merge")}
    
  ## join the emission rate data
    region_count_emissions <- region_count_emissions %>% 
      left_join(manure_all_em, by=c("iso3c", "animal", "system", "product")) %>%
      rowwise() %>%
      mutate(activity = current * kg_n_yr) %>%
      mutate(managed_em = activity * managed_manure_prop * N_to_N20_managed) %>%
      mutate(managed_spread_em = activity * managed_manure_prop * (1 - loss_N_managed_prop) * N20_to_N_spreadonfields) %>%
      mutate(spread_em = activity * spread_on_field_prop * N20_to_N_spreadonfields) %>%
      mutate(leftonfield_em = activity * left_on_field_prop * N20_to_N_leftonfields) %>%
      mutate(total_manure_tonnesCO2_em =  0.001 * N20_to_CO2 * (managed_em + managed_spread_em + spread_em + leftonfield_em))
    
    if(dim(region_count_emissions)[1] != 9331200){cat("error in count after merge")}      
    
n2o_emissions <- region_count_emissions %>%
  group_by(iso3c, animal, system, product) %>%
  dplyr::summarize(animals = sum(current, na.rm=TRUE),
                   kg_n_yr = mean(kg_n_yr, na.rm=TRUE),
                   managed_em = sum(managed_em, na.rm=TRUE),
                   managed_spread_em = sum(managed_spread_em, na.rm=TRUE),
                   spread_em = sum(spread_em, na.rm=TRUE),
                   leftonfield_em = sum(leftonfield_em, na.rm=TRUE),
                   total_manure_CO2_em = sum(total_manure_tonnesCO2_em, na.rm=TRUE)) %>%
  filter(!is.na(iso3c))
    
   write_csv(n2o_emissions, here(paste0("animal_farm/ghg/data/manure_n2o_emissions/n2o_", animal_name, "_", system_name,"_", product_name, ".csv", sep = "")))
   
   map_em <- region_count_emissions %>% 
      select(x,y, total_manure_tonnesCO2_em) %>% 
      rasterFromXYZ(crs = food_crs)
    
    writeRaster(map_em, filename = paste(prep, "animal_farm/ghg/", animal_name, "_", system_name, "_", product_name,"_n2o_manure.tif", sep = ""), format = "GTiff", overwrite = TRUE)

   
}

```

### Checking data

```{r}

#compare global animals for all emissions:
lop <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_leftonpasture_2-18-2021.csv") %>%
  filter(Element == "Emissions (CO2eq) (Manure on pasture)")
  
managed <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_managed_2-18-2021.csv") %>%
  filter(Element == "Emissions (CO2eq) from N2O (Manure management)")

sof <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_appliedtosoils_2-18-2021.csv") %>%
  filter(Element == "Emissions (CO2eq) (Manure applied)")
  
emissions <- rbind(lop, managed) %>%
  rbind(sof) %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value) %>%
  group_by(Item) %>%
  summarize(CO2eq = sum(Value, na.rm=TRUE) * 300/310 *1000)
  
files <- list.files(path = here("animal_farm/ghg/data/manure_n2o_emissions"), pattern = "*.csv", full.names = T)
n2o_our_ests <- sapply(files, read_csv, simplify=FALSE) %>% 
bind_rows(.id = "id")

n2o_our_ests %>% group_by(animal, product) %>%
  summarize(sum(total_manure_CO2_em))

## compare separate inputs
lop <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_leftonpasture_2-18-2021.csv") %>%
  filter(Element == "Emissions (N2O) (Manure on pasture)") %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value)

lop %>%
  group_by(Item) %>%
  summarize(lop_N20_fao = sum(Value, na.rm=TRUE) * 1000000)

files <- list.files(path = here("animal_farm/ghg/data/manure_n2o_emissions"), pattern = "*.csv", full.names = T)
n2o_our_ests <- sapply(files, read_csv, simplify=FALSE) %>% 
bind_rows(.id = "id")

n2o_our_ests %>% 
  group_by(animal, product) %>%
  summarize(lop_us = sum(leftonfield_em, na.rm=TRUE))


### managed
managed <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_managed_2-18-2021.csv") %>%
  filter(Element == "Emissions (N2O) (Manure management)") %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value)

managed %>%
  group_by(Item) %>%
  summarize(managed_N20_fao = sum(Value, na.rm=TRUE) * 1000000)

n2o_our_ests %>% 
  group_by(animal, product) %>%
  summarize(managed_us = sum(managed_em, na.rm=TRUE))


### spread on soils
sof <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_appliedtosoils_2-18-2021.csv") %>%
  filter(Element == "Emissions (N2O) (Manure applied)") %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value)

sof %>%
  group_by(Item) %>%
  summarize(spread_N20_fao = sum(Value, na.rm=TRUE) * 1000000)

n2o_our_ests %>% 
  group_by(animal, product) %>%
  summarize(spread_us = sum(spread_em, na.rm=TRUE))

n2o_our_ests %>% 
  group_by(animal, product) %>%
  summarize(spread_us = sum(managed_spread_em, na.rm=TRUE))


buffaloes_us = sum(c(78.4, 75.2, 73.4))
buffaloes_fao = sum(c(140.7, 35.2, 41.8))

chickens_eggs_us = sum(c(47.1/2, 0, 0/2, 40.7, 28.1/2, 0/2, 6.5, 33.2))
chickens_eggs_fao = sum(c(62.3, 19.1, 53.7))

chickens_eggs_us = sum(c(47.1/2, 0, 0/2, 69, 28.1/2, 0/2, 0, 99.4))
chickens_eggs_fao = sum(c(85.4, 26.9, 75.7))

goats_us = sum(c(40.5, 11.9, 2.2, 20.3))
goats_fao = sum(c(336, 3.7, 11.9))

sheep_us = sum(c(77.5, 17.3, 3.3, 21.7))
sheep_fao = sum(c(321, 11.1, 27.7))

pigs_us = sum(c(0, 165.7, 6.5, 173.7))
pigs_fao = sum(c(0, 121.3, 125.5))

cows_milk_us = sum(c(182.9, 64.5, 8.8, 87.2))
cows_milk_fao = sum(c(329.8, 57.7, 122.2))

cows_meat_us = sum(c(944.2, 363.5, 13.8, 325.4))
cows_meat_fao = sum(c(1382.6, 159.7, 119.6))


## why are goats/sheep so low?
lop <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_leftonpasture_2-18-2021.csv") %>%
  filter(Element == "Stocks") %>%
  filter(Area != "China")%>%
  mutate(Area = ifelse(Area == "Netherlands Antilles (former)", "Bonaire, Sint Eustatius and Saba", Area)) %>%
  select(Area, Item, Element, Value)

sof %>%
  group_by(Item) %>%
  summarize(counts = sum(Value, na.rm=TRUE))

lop <- read_csv("/home/shares/food-systems/Food_footprint/_raw_data/FAO_data/v2021/manure_management/FAOSTAT_manure_leftonpasture_2-18-2021.csv")  %>%
  filter(Area == "Australia") %>%
  select(Element, Item, Value) %>%
  pivot_wider(names_from = Element, values_from=Value)

write_csv(lop, here("checking.csv"))

n2o_our_ests %>% 
filter(iso3c == "AUS") %>% 
  select(-id) %>% data.frame()
```

```{r}

library(plotly)
library(hrbrthemes)

## wrangle our totals
source(here("_spatial/template_raster.R"))

list_files <- list.files(file.path(prep, "animal_farm/ghg"), pattern = "n2o_manure",  full = TRUE)

stack_manure_n2o <- raster::stack(lapply(list_files, raster))
manure_n2o <- sum(stack_manure_n2o, na.rm = TRUE)
    
manure_co2eq <- manure_n2o*298

our_totals <- zonal(manure_co2eq, food_rgns_tif, fun = "sum") %>% 
  as_tibble() %>% 
  dplyr::rename(ID_0 = zone) %>% 
  left_join(food_rgns, by = "ID_0") %>% 
  rename(our_total_co2eq = sum)

```
